pyproject.toml
[project]
name = "chessbot2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "lightning>=2.6.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "python-chess>=1.999",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "rootutils>=1.0.7",
    "torch>=2.11.0",
    "tqdm>=4.67.3",
    "wandb>=0.25.1",
    "zstandard>=0.25.0",
]

[tool.ruff]
line-length = 88
preview = true
target-version = "py311"
lint.select = ["ALL", "D212", "D417"]
lint.ignore = [
    "COM", "D100", "D101", "D102", "D103", "D104", "D105", "D205", "D401", "EM","FIX",
    "FBT","S101", "S404", "S602", "PLR2004", "PLR0912", "PLR0913", "PLR0914", "PLR0915",
    "PLR0917","PLC0415", "G004", "PD901", "N802", "C901", "DTZ005", "DTZ007",
    "INP", "EXE002", "TD002", "ANN001", "ANN002", "ANN003", "ANN201",
    "ANN202", "ANN204", "CPY001", "TRY003", "PLR1702", "N803", "N806", "N812", "T201",
    "PLW1514", "PTH123", "RUF015", "RUF017", "PLR6301","ERA", "ISC001", "ANN401",
    "PLR0911", "BLE001", "PD011", "S110", "E722", "S311", "DOC201", "DOC501", "E402",
    "RUF001", "RUF003", "PLR6104"
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false
